requirements.txt
streamlit>=1.28.0
openai>=1.6.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
html2text>=2020.1.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
langchain>=0.0.350
langchain-openai>=0.0.2
